requirements.txt
opencv-python==4.5.3.56
mediapipe==0.8.9
pyalsaaudio==0.9.0